{{cookiecutter.project_name}}/pyproject.toml
[project]
name = "slidedeck"
version = "0.1.0"
description = "CLI tool for managing research slide decks"
authors = [{ name = "{{ cookiecutter.author }}" }]
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
]

[project.scripts]
slidedeck = "slidedeck.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/slidedeck"]

[tool.ruff]
line-length = 120
